More precise and detailed information "Learn Ethical Hacking From Scratch | Udemy" by Zaid Sabih move down to it's own section.
Using "upgrade" and "dist-upgrade" is the same as "full-upgrade". Knowing this, the instruction were changed.

Rebooting after blacklisting a module is completely optional.
Compiling the repository will install the driver in the currently kernel.
Rebooting after the full-upgrade is recommended, because this might had installed a new kernel.

Added instructions to go back to Managed mode.

NetworkManager configuration section got changed to a more user-friendly approach.

Added Troubleshooting section in order to help users.